pyproject.toml
[tool.poetry]
name = "ccom-logger"
version = "0.1.0"
description = ""
authors = ["Jani Hietala <jani.hietala@vtt.fi>", "Petri Kaarmila <petri.kaarmila@vtt.fi>"]

[tool.poetry.dependencies]
python = "^3.8"
paho-mqtt = "^1.6.1"
pytz = "^2022.1"
psycopg2-binary = "^2.9.3"
pgcopy = "^1.5.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pycodestyle = "^2.8.0"
pylint = "^2.13.8"
pytest = "^7.1.2"
mypy = "^0.950"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"